+++ jest.config.js
@@ -1,4 +1,5 @@
 export default {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  transformIgnorePatterns: ['!node_modules/'],
 };
