+++ tests/units/storage.spec.ts
@@ -1,8 +1,5 @@
 import { FileObject } from '../../src/@types/fileObject'
-import { Storage } from '../../src/components/storage'
-import { ArweaveStorage } from '../../src/components/storage/arweave'
-import { IpfsStorage } from '../../src/components/storage/ipfs'
-import { UrlStorage } from '../../src/components/storage/url'
+import { Storage, UrlStorage, ArweaveStorage, IpfsStorage } from '../../src/components/storage'
 
 describe('URL Storage tests', () => {
   let file: FileObject = {
@@ -25,7 +22,7 @@ describe('URL Storage tests', () => {
     expect(storage).toBeInstanceOf(UrlStorage)
   })
   it('URL validation passes', () => {
-    expect(storage.validate()).toBe([true, ''])
+    expect(storage.validate()).toStrictEqual([true, ''])
   })
   it('URL validation fails on missing URL', () => {
     file = {
@@ -39,7 +36,7 @@ describe('URL Storage tests', () => {
       ]
     }
     storage = Storage.getStorageClass(file)
-    expect(storage.validate()).toBe([false, 'URL or method are missing!'])
+    expect(storage.validate()).toStrictEqual([false, 'URL or method are missing!'])
     file = {
       type: 'url',
       url: 'http://someUrl.com/file.json',
@@ -51,11 +48,12 @@ describe('URL Storage tests', () => {
       ]
     }
     storage = Storage.getStorageClass(file)
-    expect(storage.validate()).toBe([false, 'URL or method are missing!'])
+    expect(storage.validate()).toStrictEqual([false, 'URL or method are missing!'])
   })
   it('URL validation fails on invalid method', () => {
     file = {
       type: 'url',
+      url: 'http://someUrl.com/file.json',
       method: 'put',
       headers: [
         {
@@ -65,13 +63,13 @@ describe('URL Storage tests', () => {
       ]
     }
     storage = Storage.getStorageClass(file)
-    expect(storage.validate()).toBe([false, 'Invalid method for URL'])
+    expect(storage.validate()).toStrictEqual([false, 'Invalid method for URL'])
   })
 
   it('URL validation fails on filename', () => {
     file = {
       type: 'url',
-      url: 'http://someUrl.com/../dir/.file.json',
+      url: './../dir/file.json',
       method: 'get',
       headers: [
         {
@@ -81,7 +79,7 @@ describe('URL Storage tests', () => {
       ]
     }
     storage = Storage.getStorageClass(file)
-    expect(storage.validate()).toBe([false, 'URL looks like a file path'])
+    expect(storage.validate()).toStrictEqual([false, 'URL looks like a file path'])
   })
   it('Gets download URL', () => {
     file = {
@@ -96,7 +94,7 @@ describe('URL Storage tests', () => {
       ]
     }
     storage = Storage.getStorageClass(file)
-    expect(storage.getDownloadUrl()).toBe('http://someUrl.com/file.json')
+    expect(storage.getDownloadUrl()).toStrictEqual('http://someUrl.com/file.json')
   })
 })
 
@@ -114,21 +112,21 @@ describe('IPFS Storage tests', () => {
     expect(storage).toBeInstanceOf(IpfsStorage)
   })
   it('IPFS validation passes', () => {
-    expect(storage.validate()).toBe([true, ''])
+    expect(storage.validate()).toStrictEqual([true, ''])
   })
   it('IPFS validation fails', () => {
     file = {
       type: 'ipfs'
     }
     storage = Storage.getStorageClass(file)
-    expect(storage.validate()).toBe([false, 'Missing CID'])
+    expect(storage.validate()).toStrictEqual([false, 'Missing CID'])
   })
 })
 
 describe('Arweave Storage tests', () => {
   let file: FileObject = {
     type: 'arweave',
-    hash: '0x2563ed54abc0001bcaef'
+    transactionId: '0x2563ed54abc0001bcaef'
   }
   let storage: Storage
   beforeAll(() => {
@@ -139,13 +137,13 @@ describe('Arweave Storage tests', () => {
     expect(storage).toBeInstanceOf(ArweaveStorage)
   })
   it('Arweave validation passes', () => {
-    expect(storage.validate()).toBe([true, ''])
+    expect(storage.validate()).toStrictEqual([true, ''])
   })
   it('Arweave validation fails', () => {
     file = {
       type: 'arweave'
     }
     storage = Storage.getStorageClass(file)
-    expect(storage.validate()).toBe([false, 'Missing transaction ID'])
+    expect(storage.validate()).toStrictEqual([false, 'Missing transaction ID'])
   })
 })
